Fix the segfault
The issue came from calls to Qml objects (line_series.replace) in a thread other than main. There is now a timer in the main responsible for updating the GUI.

The event display is not impacted because the corresponding class is thread-safe.

This commit also includes minor fixes to the Rust code.